force file with bad state to be removed during checkout
When removeUnknownFiles flag is specified during checkout stage, but file is in "bad state"
(for example symlink that points to itself, which can be created with ln -s link link), osmosis
fails to complete checkout. According to boost documentation this error state means that it was impossible
to determine file attributes and whether file actually exists or not. In this case it is better to dispose this file,
as it cannot be used anyhow. (see http://www.boost.org/doc/libs/1_44_0/libs/filesystem/v3/doc/reference.html#status)